+++ doc/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+.. rubric:: Development version
+
+- Raise :exc:`ValueError` on a dtype that has zero itemsize (#37).
+- Change exception when dtype has embedded objects from :exc:`TypeError` to
+  :exc:`ValueError` for consistency
+
 .. rubric:: 1.13.0
 
 - Significant performance improvements to send code (in some cases an order of